+++ lib/i18n/tasks/command/commands/meta.rb
@@ -5,12 +5,12 @@ module I18n::Tasks
         include Command::Collection
 
         cmd :config,
-            args: '[section ...]',
+            pos:  '[section ...]',
             desc: t('i18n_tasks.cmd.desc.config')
 
         def config(opts = {})
           cfg = i18n.config_for_inspect
-          cfg = cfg.slice(*opts[:arguments]) if opts[:arguments]
+          cfg = cfg.slice(*opts[:arguments]) if opts[:arguments].present?
           cfg = cfg.to_yaml
           cfg.sub! /\A---\n/, ''
           cfg.gsub! /^([^\s-].+?:)/, Term::ANSIColor.cyan(Term::ANSIColor.bold('\1'))