+++ lib/i18n/tasks/command/options/common.rb
@@ -1,46 +1,36 @@
-require 'i18n/tasks/command/options/enum_opt'
-require 'i18n/tasks/command/options/list_opt'
 
 module I18n::Tasks
   module Command
     module Options
       module Common
         include Command::DSL
-        include Options::EnumOpt
-        include Options::ListOpt
 
-        VALID_LOCALE_RE = /\A\w[\w\-\.]*\z/i
+        arg :nostdin,
+            '-S',
+            '--nostdin',
+            t('i18n_tasks.cmd.args.desc.nostdin')
 
-        cmd_opt :nostdin, {
-            short: :S,
-            long:  :nostdin,
-            desc:  t('i18n_tasks.cmd.args.desc.nostdin'),
-            conf:  {default: false}
-        }
+        arg :confirm,
+            '-y',
+            '--confirm',
+            desc: t('i18n_tasks.cmd.args.desc.confirm')
 
-        cmd_opt :confirm, {
-            short: :y,
-            long:  :confirm,
-            desc:  t('i18n_tasks.cmd.args.desc.confirm'),
-            conf:  {default: false}
-        }
+        arg :pattern,
+            '-p',
+            '--pattern PATTERN',
+            t('i18n_tasks.cmd.args.desc.key_pattern')
 
-        cmd_opt :pattern, {
-            short: :p,
-            long:  :pattern=,
-            desc:  t('i18n_tasks.cmd.args.desc.key_pattern'),
-            conf:  {argument: true, optional: false}
-        }
+        arg :value,
+            '-v',
+            '--value VALUE',
+            t('i18n_tasks.cmd.args.desc.value')
 
-        cmd_opt :value, {
-            short: :v,
-            long:  :value=,
-            desc:  t('i18n_tasks.cmd.args.desc.value'),
-            conf:  {argument: true, optional: false}
-        }
+        def arg_or_pos!(key, opts)
+          opts[key] ||= opts[:arguments].try(:shift)
+        end
 
-        def opt_or_arg!(key, opt)
-          opt[key] ||= opt[:arguments].try(:shift)
+        def pos_or_stdin!(opts)
+          opts[:arguments].try(:shift) || $stdin.read
         end
       end
     end