+++ lib/i18n/tasks/scanners/pattern_scanner.rb
@@ -13,15 +13,17 @@ module I18n::Tasks::Scanners
       text = opts[:text] || read_file(path)
       text.scan(pattern) do |match|
         src_pos = Regexp.last_match.offset(0).first
-        key     = match_to_key(match, path)
-        next unless valid_key?(key, strict)
-        key = key + ':' if key.end_with?('.')
         location = src_location(path, text, src_pos)
-        unless exclude_line?(location[:line], path)
-          keys << [key, data: location]
-        end
+        next if exclude_line?(location[:line], path)
+        key = match_to_key(match, path, location)
+        next unless key
+        key = key + ':' if key.end_with?('.')
+        next unless valid_key?(key, strict)
+        keys << [key, data: location]
       end
       keys
+    rescue Exception => e
+      raise ::I18n::Tasks::CommandError.new("Error scanning #{path}: #{e.message}")
     end
 
     def default_pattern
@@ -38,10 +40,31 @@ module I18n::Tasks::Scanners
     # @param [MatchData] match
     # @param [String] path
     # @return [String] full absolute key name
-    def match_to_key(match, path)
+    def match_to_key(match, path, location)
       key = strip_literal(match[0])
-      key = absolutize_key(key, path) if path && key.start_with?('.')
-      key
+      absolute_key(key, path, location)
+    end
+
+    def absolute_key(key, path, location)
+      if key.start_with?('.')
+        if controller_file?(path)
+          absolutize_key(key, path, relative_roots, closest_method(location))
+        else
+          absolutize_key(key, path)
+        end
+      else
+        key
+      end
+    end
+
+    def controller_file?(path)
+      /controllers/.match(path)
+    end
+
+    def closest_method(location)
+      method = File.readlines(location[:src_path], encoding: 'UTF-8').first(location[:line_num] - 1).reverse_each.find { |x| x=~ /\bdef\b/ }
+      method &&= method.strip.sub(/^def\s*/, '').sub(/[\(\s;].*$/, '')
+      method
     end
 
     def pattern