+++ spec/commands/tree_commands_spec.rb
@@ -13,7 +13,7 @@ describe 'Tree commands' do
   context 'tree-merge' do
     trees = [{'a' => '1', 'b' => '2'}, {'a' => '-1', 'c' => '3'}]
     it trees.map(&:to_json).join(', ') do
-      merged = JSON.parse run_cmd(:tree_merge, format: 'json', arguments: trees.map(&:to_json), nostdin: true)
+      merged = JSON.parse run_cmd('tree-merge', '-fjson', '-S', *trees.map(&:to_json))
       expect(merged).to eq trees.reduce(:merge)
     end
   end
@@ -22,7 +22,7 @@ describe 'Tree commands' do
     forest  = {'a' => '1', 'b' => '2', 'c' => {'a' => '3'}}
     pattern = '{a,c.*}'
     it "-p #{pattern.inspect} #{forest.to_json}" do
-      selected = JSON.parse run_cmd(:tree_filter, format: 'json', pattern: pattern, arguments: [forest.to_json])
+      selected = JSON.parse run_cmd('tree-filter', '-fjson', '-p', pattern, forest.to_json)
       expect(selected).to eq(forest.except('b'))
     end
   end
@@ -30,7 +30,7 @@ describe 'Tree commands' do
   context 'tree-subtract' do
     trees = [{'a' => '1', 'b' => '2'}, {'a' => '-1', 'c' => '3'}]
     it trees.map(&:to_json).join(' - ') do
-      subtracted = JSON.parse run_cmd(:tree_subtract, format: 'json', arguments: trees.map(&:to_json), nostdin: true)
+      subtracted = JSON.parse run_cmd('tree-subtract', '-fjson', '-S', *trees.map(&:to_json))
       expected = {'b' => '2'}
       expect(subtracted).to eq expected
     end
@@ -41,17 +41,18 @@ describe 'Tree commands' do
       {'a' => {'b' => {'a' => '1'}}}
     end
 
+    def rename_key(from, to)
+      JSON.parse run_cmd('tree-rename-key', '-fjson', '-k', from, '-n', to, forest.to_json)
+    end
+
     it 'renames root node' do
-      renamed = JSON.parse run_cmd(:tree_rename_key, key: 'a', name: 'x', format: 'json', arguments: [forest.to_json])
-      expect(renamed).to eq(forest.tap { |f| f['x'] = f.delete('a') })
+      expect(rename_key('a', 'x')).to eq(forest.tap { |f| f['x'] = f.delete('a') })
     end
     it 'renames node' do
-      renamed = JSON.parse run_cmd(:tree_rename_key, key: 'a.b', name: 'x', format: 'json', arguments: [forest.to_json])
-      expect(renamed).to eq(forest.tap { |f| f['a']['x'] = f['a'].delete('b') })
+      expect(rename_key('a.b', 'x')).to eq(forest.tap { |f| f['a']['x'] = f['a'].delete('b') })
     end
     it 'renames leaf' do
-      renamed = JSON.parse run_cmd(:tree_rename_key, key: 'a.b.a', name: 'x', format: 'json', arguments: [forest.to_json])
-      expect(renamed).to eq(forest.tap { |f| f['a']['b']['x'] = f['a']['b'].delete('a') })
+      expect(rename_key('a.b.a', 'x')).to eq(forest.tap { |f| f['a']['b']['x'] = f['a']['b'].delete('a') })
     end
   end
 
@@ -61,7 +62,7 @@ describe 'Tree commands' do
     end
 
     it 'converts to keys' do
-      keys = run_cmd(:tree_convert, from: 'json', to: 'keys', arguments: [forest.to_json]).split("\n")
+      keys = run_cmd('tree-convert', '-fjson', '-tkeys', forest.to_json).split("\n")
       expect(keys.sort).to eq(['a.b.a', 'x'].sort)
     end
   end