+++ spec/i18n_tasks_spec.rb
@@ -1,16 +1,38 @@
 # coding: utf-8
 require 'spec_helper'
 require 'fileutils'
+require 'open3'
 
+# Integration tests
 describe 'i18n-tasks' do
-  delegate :run_cmd, :i18n_task, :in_test_app_dir, to: :TestCodebase
+  delegate :run_cmd, :run_cmd_capture_stderr, :i18n_task, :in_test_app_dir, to: :TestCodebase
+
+  describe 'bin/i18n-tasks' do
+    it 'shows help when invoked with no arguments, shows version on --version' do
+      # These bin/i18n-tasks tests are executed in parallel for performance
+      [
+          proc {
+            out, err, status = Open3.capture3('bin/i18n-tasks')
+            expect(status).to be_success
+            expect(out).to be_empty
+            expect(err).to start_with('Usage: i18n-tasks [command] [options]')
+            expect(err).to include('Available commands', 'add-missing')
+          },
+          proc {
+            expect(%x[bin/i18n-tasks --version].chomp).to eq(I18n::Tasks::VERSION)
+          }
+      ].map { |test| Thread.start(&test) }.each(&:join)
+    end
+  end
+
+  # Tests execute i18n-tasks using I18n::Tasks::CLI directly, via #run_cmd(task, *arguments).
+  # This avoid launching a process for each command.
 
   describe 'health' do
     it 'outputs stats' do
       t     = i18n_task
-      stats = in_test_app_dir { t.forest_stats(t.data_forest t.locales) }
-      out   = capture_stderr { run_cmd :health }
-      stats.values.each do |v|
+      out   = run_cmd_capture_stderr('health')
+      in_test_app_dir { t.forest_stats(t.data_forest t.locales) }.values.each do |v|
         expect(out).to include(v.to_s)
       end
     end
@@ -18,65 +40,72 @@ describe 'i18n-tasks' do
 
   describe 'missing' do
     let (:expected_missing_keys) {
-      %w( en.used_but_missing.key en.relative.index.missing
+      %w( en.used_but_missing.key
+          en.relative.index.missing
           es.missing_in_es.a
           en.present_in_es_but_not_en.a
-          en.hash.pattern_missing.a en.hash.pattern_missing.b
-          en.missing_symbol_key en.missing_symbol.key_two en.missing_symbol.key_three
-          es.missing_in_es_plural_1.a es.missing_in_es_plural_2.a
+          en.hash.pattern_missing.a
+          en.hash.pattern_missing.b
+          en.missing_symbol_key
+          en.missing_symbol.key_two
+          en.missing_symbol.key_three
+          es.missing_in_es_plural_1.a
+          es.missing_in_es_plural_2.a
           en.missing-key-with-a-dash.key
           en.missing-key-question?.key
-          en.fn_comment en.only_in_es
+          en.fn_comment
+          en.only_in_es
+          en.events.show.success
         )
     }
     it 'detects missing' do
-      capture_stderr do
-        expect(run_cmd :missing).to be_i18n_keys expected_missing_keys
-        es_keys = expected_missing_keys.grep(/^es\./)
-        # locale argument
-        expect(run_cmd :missing, locales: %w(es)).to be_i18n_keys es_keys
-        expect(run_cmd :missing, arguments: %w(es)).to be_i18n_keys es_keys
-      end
+      es_keys = expected_missing_keys.grep(/^es\./)
+      expect(run_cmd 'missing').to be_i18n_keys expected_missing_keys
+      # locale argument
+      expect(run_cmd 'missing', '-les').to be_i18n_keys es_keys
+      expect(run_cmd 'missing', 'es').to be_i18n_keys es_keys
     end
   end
 
   describe 'eq_base' do
-    it 'detects eq_base' do
-      capture_stderr do
-        expect(run_cmd :eq_base).to be_i18n_keys %w(es.same_in_es.a)
-      end
+    it 'detects eq-base' do
+      expect(run_cmd 'eq-base').to be_i18n_keys %w(es.same_in_es.a)
     end
   end
 
-  let(:expected_unused_keys) { %w(unused.a unused.numeric unused.plural).map { |k| %w(en es).map { |l| "#{l}.#{k}" } }.reduce(:+) }
-  let(:expected_unused_keys_strict) { expected_unused_keys + %w(hash.pattern.a hash.pattern2.a).map { |k| %w(en es).map { |l| "#{l}.#{k}" } }.reduce(:+) }
+  let(:expected_unused_keys) do
+    %w(unused.a unused.numeric unused.plural).map do |k|
+      %w(en es).map { |l| "#{l}.#{k}" }
+    end.reduce(:+)
+  end
+
+  let(:expected_unused_keys_strict) do
+    expected_unused_keys + %w(hash.pattern.a hash.pattern2.a).map do |k|
+      %w(en es).map { |l| "#{l}.#{k}" }
+    end.reduce(:+)
+  end
+
   describe 'unused' do
     it 'detects unused' do
-      capture_stderr do
-        expect(run_cmd :unused).to be_i18n_keys expected_unused_keys
-      end
+      expect(run_cmd 'unused').to be_i18n_keys expected_unused_keys
     end
 
     it 'detects unused (--strict)' do
-      capture_stderr do
-        expect(run_cmd :unused, strict: true).to be_i18n_keys expected_unused_keys_strict
-      end
+      expect(run_cmd 'unused', '--strict').to be_i18n_keys expected_unused_keys_strict
     end
   end
 
   describe 'remove_unused' do
     it 'removes unused' do
       in_test_app_dir do
-        t = i18n_task
+        t      = i18n_task
         unused = expected_unused_keys.map { |k| ::I18n::Tasks::SplitKey.split_key(k, 2)[1] }
         unused.each do |key|
           expect(t.key_value?(key, :en)).to be true
           expect(t.key_value?(key, :es)).to be true
         end
         ENV['CONFIRM'] = '1'
-        capture_stderr {
-          run_cmd :remove_unused
-        }
+        run_cmd 'remove-unused'
         t.data.reload
         unused.each do |key|
           expect(t.key_value?(key, :en)).to be false
@@ -89,7 +118,7 @@ describe 'i18n-tasks' do
   describe 'normalize' do
     it 'sorts the keys' do
       in_test_app_dir do
-        run_cmd :normalize
+        run_cmd 'normalize'
         en_yml_data = i18n_task.data.reload['en'].select_keys { |_k, node|
           node.data[:path] == 'config/locales/en.yml'
         }
@@ -105,7 +134,7 @@ describe 'i18n-tasks' do
     it 'moves keys to the corresponding files as per data.write' do
       in_test_app_dir {
         expect(File).to_not exist 'config/locales/devise.en.yml'
-        run_cmd :normalize, pattern_router: true
+        run_cmd 'normalize', '--pattern_router'
         expect(YAML.load_file('config/locales/devise.en.yml')['en']['devise']['a']).to eq 'EN_TEXT'
       }
     end
@@ -114,7 +143,7 @@ describe 'i18n-tasks' do
   describe 'xlsx_report' do
     it 'saves' do
       in_test_app_dir {
-        capture_stderr { run_cmd :xlsx_report }
+        run_cmd 'xlsx-report'
         expect(File).to exist 'tmp/i18n-report.xlsx'
         FileUtils.cp('tmp/i18n-report.xlsx', '..')
       }
@@ -127,7 +156,7 @@ describe 'i18n-tasks' do
       in_test_app_dir {
         expect(YAML.load_file('config/locales/en.yml')['en']['used_but_missing']).to be_nil
       }
-      run_cmd :add_missing, locales: 'base'
+      run_cmd 'add-missing', 'base'
       in_test_app_dir {
         expect(YAML.load_file('config/locales/en.yml')['en']['used_but_missing']['key']).to eq I18n.t('i18n_tasks.common.key')
         expect(YAML.load_file('config/locales/en.yml')['en']['present_in_es_but_not_en']['a']).to eq 'ES_TEXT'
@@ -138,7 +167,7 @@ describe 'i18n-tasks' do
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']).to be_nil
       }
-      run_cmd :add_missing, locales: 'es'
+      run_cmd 'add-missing', 'es'
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']['a']).to eq 'EN_TEXT'
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es_plural_1']['a']['one']).to eq 'EN_TEXT'
@@ -149,8 +178,8 @@ describe 'i18n-tasks' do
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']).to be_nil
       }
-      run_cmd :normalize, pattern_router: true
-      run_cmd :add_missing, locales: 'all', value: 'TRME'
+      run_cmd 'normalize', '--pattern_router'
+      run_cmd 'add-missing', '-v', 'TRME'
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']['a']).to eq 'TRME'
         expect(YAML.load_file('config/locales/devise.es.yml')['es']['devise']['a']).to eq 'ES_TEXT'
@@ -162,7 +191,7 @@ describe 'i18n-tasks' do
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']).to be_nil
       }
-      run_cmd :add_missing, locales: 'all', value: 'TRME %{value}'
+      run_cmd 'add-missing', '-v', 'TRME %{value}'
       in_test_app_dir {
         expect(YAML.load_file('config/locales/es.yml')['es']['missing_in_es']['a']).to eq 'TRME EN_TEXT'
         expect(YAML.load_file('config/locales/en.yml')['en']['present_in_es_but_not_en']['a']).to eq 'TRME ES_TEXT'
@@ -172,7 +201,7 @@ describe 'i18n-tasks' do
 
   describe 'config' do
     it 'prints config' do
-      expect(YAML.load(Term::ANSIColor.uncolor(run_cmd :config))).to(
+      expect(YAML.load(Term::ANSIColor.uncolor(run_cmd 'config'))).to(
           eq(in_test_app_dir { i18n_task.config_for_inspect })
       )
     end
@@ -180,20 +209,17 @@ describe 'i18n-tasks' do
 
   describe 'find' do
     it 'prints usages' do
-      capture_stderr do
-        result = Term::ANSIColor.uncolor(run_cmd :find, arguments: ['used.*'])
-        expect(result).to eq(<<-TXT)
+      result = Term::ANSIColor.uncolor(run_cmd 'find', 'used.*')
+      expect(result).to eq(<<-TXT)
 used.a 2
   app/views/usages.html.slim:1 p = t 'used.a'
   app/views/usages.html.slim:2 b = t 'used.a'
-        TXT
-      end
+      TXT
     end
   end
 
-
   # --- setup ---
-  BENCH_KEYS = 10
+  BENCH_KEYS = ENV['BENCH_KEYS'].to_i
   before(:each) do
     gen_data = ->(v) {
       v_num = v.chars.map(&:ord).join('').to_i
@@ -225,21 +251,25 @@ used.a 2
           'devise'                 => {'a' => v},
           'scoped'                 => {'x' => v},
           'very'                   => {'scoped' => {'x' => v}},
-          'used'                   => {'a' => v}
+          'used'                   => {'a' => v},
+          'latin_extra'            => {'çüéö' => v},
+          'not_a_comment'          => v
       }.tap { |r|
-        gen = r["bench"] = {}
-        BENCH_KEYS.times { |i| gen["key#{i}"] = v }
+        if BENCH_KEYS > 0
+          gen = r['bench'] = {}
+          BENCH_KEYS.times { |i| gen["key#{i}"] = v }
+        end
       }
     }
 
-    en_data                            = gen_data.('EN_TEXT')
-    es_data                            = gen_data.('ES_TEXT').except(
-        'missing_in_es', 'missing_in_es_plural_1', 'missing_in_es_plural_2')
-    es_data['same_in_es']['a']         = 'EN_TEXT'
-    es_data['blank_in_es']['a']        = ''
-    es_data['ignore_eq_base_all']['a'] = 'EN_TEXT'
-    es_data['ignore_eq_base_es']['a']  = 'EN_TEXT'
-    es_data['only_in_es'] = 1
+    en_data = gen_data.('EN_TEXT')
+    es_data = gen_data.('ES_TEXT').except('missing_in_es', 'missing_in_es_plural_1', 'missing_in_es_plural_2')
+
+    es_data['same_in_es']['a']          = 'EN_TEXT'
+    es_data['blank_in_es']['a']         = ''
+    es_data['ignore_eq_base_all']['a']  = 'EN_TEXT'
+    es_data['ignore_eq_base_es']['a']   = 'EN_TEXT'
+    es_data['only_in_es']               = 1
     es_data['present_in_es_but_not_en'] = {'a' => 'ES_TEXT'}
 
     fs = fixtures_contents.merge(