+++ spec/relative_keys_spec.rb
@@ -17,6 +17,43 @@ describe 'Relative keys' do
       end
     end
 
-  end
+    context 'relative key in controller' do
+      it 'works' do
+        key = scanner.absolutize_key(
+          '.success',
+          'app/controllers/users_controller.rb',
+          %w(app/controllers),
+          'create'
+        )
 
+        expect(key).to eq('users.create.success')
+      end
+
+      context 'multiple words in controller name' do
+        it 'works' do
+          key = scanner.absolutize_key(
+            '.success',
+            'app/controllers/admin_users_controller.rb',
+            %w(app/controllers),
+            'create'
+          )
+
+          expect(key).to eq('admin_users.create.success')
+        end
+      end
+
+      context 'nested in module' do
+        it 'works' do
+          key = scanner.absolutize_key(
+            '.success',
+            'app/controllers/nested/users_controller.rb',
+            %w(app/controllers),
+            'create'
+          )
+
+          expect(key).to eq('nested.users.create.success')
+        end
+      end
+    end
+  end
 end